V4 Single symbols to stream of symbols
Removed rendering single symbols and renders streams instead. Has some bugs, like if a stream is too long with a huge font it will go on top of it's self when looping from bottom to top.  Also in a stream each symbol can have a different changing speed (interval). not sure if good or bad.